requirements.txt
absl-py==0.9.0
astor==0.8.1
attrs==19.3.0
Click==7.0
Flask==1.1.1
gast==0.3.2
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
importlib-metadata==1.4.0
itsdangerous==1.1.0
Jinja2==2.10.3
kafka-python==1.4.7
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
more-itertools==8.1.0
msgpack-python==0.5.6
numpy==1.18.1
opencv-python==4.1.2.30
packaging==20.0
Pillow==7.0.0
pkg-resources==0.0.0
pluggy==0.13.1
protobuf==3.11.2
py==1.8.1
pyparsing==2.4.6
pytest==5.3.4
python-memcached==1.59
six==1.14.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
wcwidth==0.1.8
Werkzeug==0.16.0
wrapt==1.11.2
zipp==2.0.0
